% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_base_theme}
\alias{.acl_base_theme}
\title{Build the standard ALSCL ggplot theme layer}
\usage{
.acl_base_theme(
  font_family = NULL,
  title_size = NULL,
  axis_title_size = NULL,
  axis_text_size = NULL,
  strip_text_size = NULL,
  legend_text_size = NULL,
  base_theme = NULL,
  title_hjust = NULL
)
}
\description{
Build the standard ALSCL ggplot theme layer
}
\keyword{internal}
