% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridges.R
\name{.acl_ridges_palette}
\alias{.acl_ridges_palette}
\title{Build fill scale for ridgeline plots}
\usage{
.acl_ridges_palette(palette, n)
}
\arguments{
\item{palette}{Character palette name or character vector of colors.}

\item{n}{Integer. Number of colors needed.}
}
\value{
A ggplot2 scale_fill layer.
}
\description{
Build fill scale for ridgeline plots
}
\keyword{internal}
