% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_scale_x}
\alias{.acl_scale_x}
\title{Build x-axis scale with breaks and expand}
\usage{
.acl_scale_x(x_breaks = NULL, n_breaks = 10)
}
\description{
Build x-axis scale with breaks and expand
}
\keyword{internal}
