% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_title}
\alias{.acl_title}
\title{Get a title from the theme, with model name substitution}
\usage{
.acl_title(key, m1 = NULL, m2 = NULL)
}
\arguments{
\item{key}{Character. Title key.}

\item{m1}{Optional model 1 name for substitution.}

\item{m2}{Optional model 2 name for substitution.}
}
\value{
Character string.
}
\description{
Get a title from the theme, with model name substitution
}
\keyword{internal}
