% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_acl.R
\name{unload_acl}
\alias{unload_acl}
\title{Unload a model dynamic library}
\usage{
unload_acl(dll_path)
}
\arguments{
\item{dll_path}{Path to the library. Existing TMB objects must no longer be used.}
}
\description{
Unload a model dynamic library
}
\keyword{internal}
