% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YTF.R
\name{YTF}
\alias{YTF}
\title{YTF(yellowtail flounder) data list}
\format{
A list with 26 elements:
\describe{
\item{nyear}{Number of time steps used for model diagnostic.}
\item{nlen}{Number of length bins.}
\item{nage}{number of age classes.}
\item{len_mid}{the midpoints of length bins.}
\item{len_border}{the border values of length bins.}
\item{len_lower}{lower bounds of length bins (cm).}
\item{len_upper}{upper bounds of length bins (cm).}
\item{vbk}{growth rate parameter.}
\item{Linf}{asymptotic length parameter.}
\item{t0}{growth parameter.}
\item{rec.age}{age class of recruitment.}
\item{init_Z}{initial mortality to determine age structure in initial time step.}
\item{cv_L}{the coefficient of variance in length at age at the beginning year.}
\item{cv_inc}{the coefficient of variance in growth increment.}
\item{M}{natural mortality.}
\item{std_logR}{standard deviation of recruitment.}
\item{std_logN0}{standard deviation of initial number at age.}
\item{std_SN}{survey measurement error.}
\item{a}{parameter to define length–weight relationship.}
\item{b}{parameter to define length–weight relationship.}
\item{alpha}{Beverton-Holt model parameter.}
\item{beta}{Beverton-Holt model parameter.}
\item{mat_L50}{length at which 50\% of individuals are mature.}
\item{mat_L95}{length at which 95\% of individuals are mature.}
\item{q_surv_L50}{survey catchability at length for 50\% catchability.}
\item{q_surv_L95}{survey catchability at length for 95\% catchability.}

}
}
\source{
Parameter values are specified based on empirical data of yellowtail flounder on the grand Bank off Newfoundland .
}
\usage{
YTF
}
\description{
Parameter values are specified based on empirical data of yellowtail flounder on the grand Bank off Newfoundland
}
\details{
This is a parameter list, not a set of survey observation tables.
Use \code{YTF_example} for reproducible, fit-ready simulated data.
}
\references{
Zhang, F. and Cadigan, N. G. (2022). Fish and Fisheries 23,
1121--1135. \doi{10.1111/faf.12673}.
}
