% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{acl_get_theme}
\alias{acl_get_theme}
\title{Internal: Build ggplot2 theme object from acl_theme()}
\usage{
acl_get_theme()
}
\value{
A list of ggplot2 theme layers.
}
\description{
This is the function called by plot functions via
\code{get("acl_get_theme", envir = asNamespace("ACL"))()}.
}
\keyword{internal}
