% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_acl.R
\name{unload_acl}
\alias{unload_acl}
\title{Unload the ACL Dynamic Library}
\usage{
unload_acl(dll_path)
}
\arguments{
\item{dll_path}{Path to the shared library file.}
}
\description{
Unload the ACL Dynamic Library
}
\keyword{internal}
